requirement.txt
absl-py==0.8.0
asn1crypto==0.24.0
astor==0.8.0
astunparse==1.6.3
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.8.2
bigzenodo==0.1.1
bleach==3.1.0
blis==0.4.1
Bottleneck==1.2.1
bytecode==0.13.0
cachetools==4.1.0
catalogue==0.2.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.1.1
cryptography==2.3
cycler==0.10.0
cymem==2.0.3
dataclasses==0.7
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
enum34==1.1.6
fastai==1.0.60
fastcore==1.4.0
fastprogress==0.2.1
fastscript==0.1.2
ffmpeg-python==0.2.0
filelock==3.4.1
future==0.18.2
gast==0.3.3
gdown==3.10.1
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.31.0
h5py==2.10.0
huggingface-hub==0.4.0
idna==2.6
imageio==2.15.0
importlib-metadata==1.3.0
ipykernel==5.1.2
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyter-http-over-ws==0.0.6
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
Markdown==3.1.1
markdown2==2.4.8
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==8.0.2
munch==2.5.0
murmurhash==1.0.2
nbconvert==5.6.0
nbdev==0.2.3
nbformat==4.4.0
networkx==2.5.1
notebook==6.0.1
numexpr==2.7.0
numpy==1.19.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-contrib-python==3.4.1.15
opencv-python==3.4.1.15
opt-einsum==3.0.1
packaging==21.3
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
patsy==0.5.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.2
plac==1.1.3
preshed==3.0.2
pretrainedmodels==0.7.4
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pycrypto==2.6.1
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
PySocks==1.7.1
pytesseract==0.3.5
python-dateutil==2.8.0
pytorch-lightning==0.9.0
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.2
pyzmq==18.1.0
qtconsole==4.5.4
QtPy==1.9.0
regex==2022.9.13
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.6
sacremoses==0.0.53
scenedetect==0.5.1.1
scikit-image==0.17.2
scikit-learn==0.22.1
scikit-posthocs==0.6.0
scipy==1.4.1
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.5.0
setuptools==58.0.4
Shapely==1.6.4
six==1.16.0
sklearn==0.0
soupsieve==1.9.5
spacy==2.2.3
srsly==1.0.1
statsmodels==0.12.2
tb-nightly==1.15.0a20190806
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019080601
tf-slim==1.1.0
thinc==7.3.1
tifffile==2020.9.3
timm==0.4.12
tokenizers==0.12.1
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.2
transformers==4.18.0
typing==3.7.4.1
typing_extensions==4.1.1
urllib3==1.25.7
wasabi==0.5.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.5
wheel==0.37.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
